leam-base.joda-time.mid-1578.idx-173908.5.mutant

+++ joda-time/src/main/java/org/joda/time/field/PreciseDurationField.java
@@ -61,9 +61,8 @@ public class PreciseDurationField extends BaseDurationField {
      * @return the unit size of this field, in milliseconds
      */
     @Override
-    public final long getUnitMillis() {
-        return iUnitMillis;
-    }
+        public final long getUnitMillis() {
+    return (long)iUnitMillis;    }
 
     //------------------------------------------------------------------------
     /**
